prompt user to proceed with install

+++ bin/install.sh
@@ -24,7 +24,26 @@ export APIGEE_HOME=$(pwd)
 # initalize other variables
 source ${APIGEE_HOME}/bin/vars.sh
 
-# step 1. install cert manager
+# step 1: prepare the kustomize files
+
+# a: generate kustomize manifests for the controller
+. ${APIGEE_HOME}/bin/generateControllerKustomize.sh
+
+# b: generate kustomize manifests for the org
+. ${APIGEE_HOME}/bin/generateOrgKustomize.sh
+
+# c: generate env kustomize manifests from templates
+. ${APIGEE_HOME}/bin/generateEnvKustomize.sh
+
+echo "Kustomize manifest files generated. "
+read -p "Do you wish to process with the installation? Select No if you wish to change default Kustomize settings:  " -n 1 -r
+echo  ""
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    exit 1
+fi
+
+# step 2. install cert manager
 helm repo add jetstack https://charts.jetstack.io && helm repo update
 
 # if using a custom repo, also add --set image.repository=gcr.io/myproject/cert-manager
@@ -35,63 +54,59 @@ else
   helm install cert-manager jetstack/cert-manager --namespace cert-manager --create-namespace --version v1.5.2 --set installCRDs=true --set nodeSelector."cloud\.google\.com/gke-nodepool"=apigee-runtime && kubectl wait deployments/cert-manager -n cert-manager --for condition=available --timeout 60s
 fi
 
-# step 2: install asm
+# step 3: install asm
 # These instructions are for GKE. If installing on another platform, see here: https://cloud.google.com/service-mesh/docs/unified-install/install#amazon-eks
-curl https://storage.googleapis.com/csm-artifacts/asm/asmcli_1.12 > ${APIGEE_HOME}/bin/asmcli
 
-chmod +x ${APIGEE_HOME}/bin/asmcli
+if [[ -z $SKIP_ASM ]]; then
+    curl https://storage.googleapis.com/csm-artifacts/asm/asmcli_1.12 > ${APIGEE_HOME}/bin/asmcli
 
-. ${APIGEE_HOME}/bin/asmcli install \
-  --project_id ${PROJECT_ID} \
-  --cluster_name ${CLUSTER_NAME} \
-  --cluster_location ${CLUSTER_REGION} \
-  --fleet_id ${PROJECT_ID} \
-  --enable_all \
-  --ca mesh_ca
+    chmod +x ${APIGEE_HOME}/bin/asmcli
 
-# step 3: install CA cert in primary region only
+    . ${APIGEE_HOME}/bin/asmcli install \
+    --project_id ${PROJECT_ID} \
+    --cluster_name ${CLUSTER_NAME} \
+    --cluster_location ${CLUSTER_REGION} \
+    --fleet_id ${PROJECT_ID} \
+    --enable_all \
+    --ca mesh_ca
+
+    # clean up asmcli
+    rm ${APIGEE_HOME}/bin/asmcli
+fi
+
+# step 4: install CA cert in primary region only
 kubectl apply -f ${APIGEE_HOME}/cluster/apigee-selfsigned-clusterissuer.yaml && kubectl wait clusterissuer/selfsigned --for condition=ready --timeout=60s
 
 kubectl apply -f ${APIGEE_HOME}/primary/apigee-ca-certificate.yaml && kubectl wait certificates/apigee-ca -n cert-manager --for condition=ready --timeout=60s
 
-# step 4: install crds
+# step 5: install crds
 kubectl create -f ${APIGEE_HOME}/cluster/crds && kubectl wait crd/apigeeenvironments.apigee.cloud.google.com --for condition=established --timeout=60s
 
-# step 5: create cluster resources
+# step 6: create cluster resources
 kubectl apply -f ${APIGEE_HOME}/cluster
 
-# step 6: install apigee controller
-. ${APIGEE_HOME}/bin/generateControllerKustomize.sh
+# step 7: install apigee controller
 kubectl apply -k ${APIGEE_HOME}/overlays/controller && kubectl wait deployments/apigee-controller-manager --for condition=available -n apigee-system --timeout=60s
-
-# step 7: generate kustomize manifests from templates
-. ${APIGEE_HOME}/bin/generateOrgKustomize.sh
-# inspect the kustomization.yaml file in ./overlays/${INSTANCE_ID}. Enable or
-# disable features as necessary
 
 # step 8: install apigee runtime instance (datastore, telemetry, redis and org)
 kubectl apply -k ${APIGEE_HOME}/overlays/${INSTANCE_ID} && kubectl wait apigeeorganizations/${ORG} -n apigee --for=jsonpath='{.status.state}'=running --timeout 300s
 
-# step 9: generate env kustomize manifests from templates
-. ${APIGEE_HOME}/bin/generateEnvKustomize.sh
-
-# step 10: install the apigee environment
+# step 9: install the apigee environment
 kubectl apply -k ${APIGEE_HOME}/overlays/${INSTANCE_ID}/environments/${ENV_NAME}
 
-# step 11: Enable Apigee envoyfilters for ASM
+# step 10: Enable Apigee envoyfilters for ASM
 kubectl apply -k ${APIGEE_HOME}/overlays/envoyfilters
 
-# clean up asmcli
-rm ${APIGEE_HOME}/bin/asmcli
-
-# step 12: This step associates GSA with KSA and adds workloadIdentityUser role to the GSA. If not using workload identity, skip the step
+# step 11: This step associates GSA with KSA and adds workloadIdentityUser role to the GSA. If not using workload identity, skip the step
 . ${APIGEE_HOME}/bin/workloadIdentity.sh
 
-# step 13: Wait for environments to be ready
+# step 12: Wait for environments to be ready
 echo "Waiting for Apigee Environment "${ENV_NAME}" to be ready"
 kubectl wait apigeeenvironments/${ENV} -n apigee --for=jsonpath='{.status.state}'=running --timeout 120s
 
 echo "Installation completed successfully!"
+
 EXTERNAL_IP=$(kubectl get svc -n apigee apigee-istio-ingressgateway --output jsonpath='{.status.loadBalancer.ingress[0].ip}' | xargs)
 EXTERNAL_HOST_NAME="api."${ORG_NAME}".example"
+
 echo "If the ingress configurations were not customized, you can test your API call with this: curl -kv https://${EXTERNAL_HOST_NAME} --resolve \"${EXTERNAL_HOST_NAME}:443:${EXTERNAL_IP}\""
